AI/requirements.txt
supabase==2.3.4
torch==2.2.1
scikit-learn==1.4.1
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1 